my_blog/requirements.txt
asgiref==3.5.0
Django==3.2.8
django-send-email==0.1.0
django-taggit==2.1.0
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
